pyproject.toml
[project]
name = "payment"
version = "0.1.0"
description = "Payment service for MentholHub"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "cryptography>=45.0.5",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "faststream[kafka]>=0.5.47",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-sugar>=1.0.0",
    "pytest-xdist>=3.8.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "autopep8>=2.3.2",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "pre-commit>=4.2.0",
    "ty>=0.0.1a14",
    "uv>=0.7.21",
]
